Taking project into different direction
`vvv-scripts` is now `vvv-manage-sites`.

The scripts can still be found in the /scripts/ folder. I've added
php that can be run inside a vagrant site that will call the ruby
scripts, giving it a bit of a "wizard" feel.

To make it work really well, you could check this out into
'www/default' and modify the default site (vvv) index.php file
with a URL to this project.

Would make setting up sites in VVV a whole lot easier.